refactor: extract lastlogin logic into separate handleLastLogin function
